added new event system and new network class. kicked out old network code. fixed calculationg path costs.
git-svn-id: svn://maxr.org/MAX/trunk@975 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ mjobs.h
@@ -66,7 +66,7 @@ public:
   void Release(void);
   bool CalcPath(void);
   int CalcDest(int x,int y);
-  bool AddPoint(int x, int y, int m, sPathCalc *p);
+  bool AddPoint(int x, int y, float m, sPathCalc *p);
   bool CreateNextPath(void);
   bool CheckPossiblePoint(int x,int y);
   bool CheckPointNotBlocked(int x,int y);
